fixed f32

+++ src/types.rs
@@ -1,5 +1,5 @@
 use cgmath::{Vector3, Quaternion};
 
-pub type Float = f64;
+pub type Float = f32;
 pub type Quat = Quaternion<Float>;
 pub type Vec3 = Vector3<Float>;
